Dinodex iteration 2 commit 2
