+++ dino_catalog/dinodex.rb
@@ -4,6 +4,7 @@ require './dino_collection'
 require './dino'
 require './filter_dinos'
 require './import_data'
+require './transform_data'
 require 'trollop'
 require 'yaml'
 require 'json'
